# swagger/awagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CRM API-DOCUMENTATION",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authorization using the Bearer scheme. Example: 'Authorization: Bearer ${token}'"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/sale": {
      "post": {
        "summary": "Create a new sale order",
        "tags": ["Sales"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "wantToUpgradeBoxes": {
                    "type": "string",
                    "example": "Yes"
                  },
                  "serviceProvider": {
                    "type": "string",
                    "example": "Spectrum"
                  },
                  "duration": {
                    "type": "string",
                    "example": "1-3 Years"
                  },
                  "securityPin": {
                    "type": "string",
                    "example": "1234"
                  },
                  "accNum": {
                    "type": "string",
                    "example": "123456789"
                  },
                  "ssn": {
                    "type": "string",
                    "example": "123-45-6789"
                  },
                  "dateOfBirth": {
                    "type": "string",
                    "format": "date",
                    "example": "1990-01-01"
                  },
                  "cardProvider": {
                    "type": "string",
                    "example": "Visa"
                  },
                  "nameOnCard": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "cardType": {
                    "type": "string",
                    "example": "Credit"
                  },
                  "cardNum": {
                    "type": "string",
                    "example": "4111111111111111"
                  },
                  "cardExp": {
                    "type": "string",
                    "example": "12/24"
                  },
                  "cvv": {
                    "type": "string",
                    "example": "123"
                  },
                  "comments": {
                    "type": "string",
                    "example": "Please handle with care"
                  },
                  "activeServices": {
                    "type": "array",
                    "items": {
                      "type": "boolean"
                    }
                  }
                },
                "required": [
                  "wantToUpgradeBoxes",
                  "serviceProvider",
                  "duration",
                  "securityPin",
                  "accNum",
                  "ssn",
                  "dateOfBirth",
                  "cardProvider",
                  "nameOnCard",
                  "cardType",
                  "cardNum",
                  "cardExp",
                  "cvv",
                  "comments",
                  "activeServices"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order placed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Order placed successfully!"
                    },
                    "order": {
                      "type": "object",
                      "properties": {
                        "clientId": {
                          "type": "integer",
                          "example": 1
                        },
                        "id": {
                          "type": "integer",
                          "example": 123
                        },
                        "Upgrade": {
                          "type": "string",
                          "example": "Yes"
                        },
                        "serviceProvider": {
                          "type": "string",
                          "example": "Spectrum"
                        },
                        "duration": {
                          "type": "string",
                          "example": "1-3 Years"
                        },
                        "services": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "securityPin": {
                          "type": "string",
                          "example": "1234"
                        },
                        "accountNum": {
                          "type": "string",
                          "example": "123456789"
                        },
                        "SSN": {
                          "type": "string",
                          "example": "123-45-6789"
                        },
                        "dob": {
                          "type": "string",
                          "format": "date",
                          "example": "1990-01-01"
                        },
                        "cardProvider": {
                          "type": "string",
                          "example": "Visa"
                        },
                        "nameOnCard": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "type": {
                          "type": "string",
                          "example": "Credit"
                        },
                        "cardNum": {
                          "type": "string",
                          "example": "4111111111111111"
                        },
                        "expiry": {
                          "type": "string",
                          "example": "12/24"
                        },
                        "CVV": {
                          "type": "string",
                          "example": "123"
                        },
                        "comment": {
                          "type": "string",
                          "example": "Please handle with care"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid service",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid service: Cable"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "An error occurred while placing the order",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "An error occurred while placing order"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sales-details": {
      "get": {
        "summary": "Retrieve sales details",
        "tags": ["Sales"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of sales details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "orderId": {
                        "type": "integer",
                        "example": 123
                      },
                      "clientName": {
                        "type": "string",
                        "example": "John Doe"
                      },
                      "agentName": {
                        "type": "string",
                        "example": "Jane Smith"
                      },
                      "Service": {
                        "type": "string",
                        "example": "Internet"
                      },
                      "serviceProvider": {
                        "type": "string",
                        "example": "Spectrum"
                      },
                      "duration": {
                        "type": "string",
                        "example": "1-3 Years"
                      },
                      "Upgradation": {
                        "type": "string",
                        "example": "Yes"
                      },
                      "securityPin": {
                        "type": "string",
                        "example": "1234"
                      },
                      "accountNum": {
                        "type": "string",
                        "example": "9876543210"
                      },
                      "SSN": {
                        "type": "string",
                        "example": "123-45-6789"
                      },
                      "cardProvider": {
                        "type": "string",
                        "example": "Visa"
                      },
                      "nameOnCard": {
                        "type": "string",
                        "example": "John Doe"
                      },
                      "CardType": {
                        "type": "string",
                        "example": "Credit"
                      },
                      "cardNum": {
                        "type": "string",
                        "example": "4111111111111111"
                      },
                      "DOB": {
                        "type": "string",
                        "format": "date",
                        "example": "1980-01-01"
                      },
                      "expiry": {
                        "type": "string",
                        "example": "12/24"
                      },
                      "CVV": {
                        "type": "string",
                        "example": "123"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User not authenticated or not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}